added Request & extracted some Router method into it
* added Request class
* refactored the whole Router to use the newly created Request class
* Removed getRequestContent() and created the "great" code inside the Request constructor
* Extracted mapRequestContent into Request and use it in the constructor
